Updated all viewmodels to use the new DbContext and related services.

+++ EveProfiteerDb/dbo/Tables/ApiKeys.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[ApiKeys] (
     [Id]       INT            IDENTITY (1, 1) NOT NULL,
-    [ApiKeyId] BIGINT         NOT NULL,
+    [ApiKeyId] INT         NOT NULL,
     [VCode]    NVARCHAR (MAX) NULL,
     [KeyType]  NVARCHAR (MAX) NULL,
     CONSTRAINT [PK_dbo.ApiKeys] PRIMARY KEY CLUSTERED ([Id] ASC)
